Lot's of updates
Added Royalties calculations
I fixed some errors in the quote sheets where the $/hr did not increase with the markup percentage.
Changed the color of "Failures" in the chart on the quote sheets to Black
Changed the background color of the totals to white to avoid confusing it with the "Failures"
Added a Template Sheet to create packets of hardware for projects that have more than six types of components
Color-coded the input on the General Sheet
Fixed error with the "Payment Fees Est." That was WAY too high

NOTE: The "Royalties" and "Payment Fees Est." are both based on the end total, so they become circular. I had to sum up "Filament, Electricity, Printer depreciation, Space Rental, Hardware/Consumables, Preparation, Post-processing, Computer Work, Failures" then do the Royalties since that will always be the more significant contributor, then estimate the Payment Fees based on all that. You'll be able to account for most everything, but it's still just an estimate. This is why it's best to round up your quote or just accept that you are going to loose a few pennies on the Payment Fees.